当たり判定みなおし。不具合調査中

+++ ゲーム科1年/00_前期/ゲームプログラミングⅢ/ゲーム数学/実習/サンプル/00_坂道の移動/完成/Project/Source/Animation/Animation.h
@@ -18,5 +18,5 @@ struct AnimationData
 void InitAnimation(AnimationData* animData);
 void UpdateAnimation(AnimationData* animData);
 void DrawAnimation(AnimationData* animData, bool isTurnX = false, bool isTurnY = false);
-void DrawAnimation(AnimationData* animData, float posX, float posY, bool isTurnX = false, bool isTurnY = false);
+void DrawAnimation(AnimationData* animData, int posX, int posY, bool isTurnX = false, bool isTurnY = false);
 void StartAnimation(AnimationData* animData, float posX, float posY, int interval, int frameNum, int width, int height, bool isLoop);
